pyproject.toml
[project]
name = "lumen"
version = "0.1.0"
description = "Monorepo for the Lumen learning-note pipeline"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi>=0.115.0",
    "openai>=1.75.0",
    "pydantic>=2.11.0",
    "sqlalchemy>=2.0.40",
    "psycopg[binary]>=3.2.0",
    "python-dotenv>=1.0.1",
    "uvicorn>=0.34.0",
]

[dependency-groups]
dev = [
    "httpx>=0.28.0",
    "pytest>=8.3.0",
]

[tool.pytest.ini_options]
pythonpath = [
    "apps/api/src",
    "packages/domain/src",
    "packages/notes-platform/src",
]
testpaths = [
    "apps/api/tests",
    "packages/domain/tests",
    "packages/notes-platform/tests",
]
